+++ src/column.component.jsx
@@ -1,8 +1,8 @@
 import styled from 'styled-components';
 
 const Column = styled.div`
-  min-width: ${props => props.width}px;
-  width: ${props => props.width}px;
+  min-width: ${props => (props.width === 'auto' ? 'auto' : `${props.width}px`)};
+  width: ${props => (props.width === 'auto' ? 'auto' : `${props.width}px`)};
   display: flex;
   padding: 0 ${props => props.theme.halfGutterWidth};
   justify-content: ${props => props.alignment};