+++ CHANGELOG.md
@@ -2,6 +2,13 @@
 
 * In general follow (https://docs.npmjs.com/getting-started/semantic-versioning) versioning.
 
+## 0.4.0
+* Added more possibilities to customize list functionality
+* Change `isSelectable` prop to `isSelectColumnVisible` (list can be selectable even if select column is not visible)
+* Add `onRowClick`, `onRowDoubleClick`, `onRowRightClick`, `onSelectAllClick` and `isAllSelected` props
+* Add 'auto' as accepted value to column width
+* Fix search when valueType is not defined on column to default to string
+
 ## 0.3.2
 * Fix `render` function index parameter